requirements.txt
langchain
langchain_community
langchain-huggingface
psycopg2-binary
pgvector
pypdf
uvicorn
fastapi

